Parsing and Fallback
The changes allow for the value (`$v` which is now `$val`) to be entered as part of the command rather than having to go to the control panel each time a user wants to use it. With that change, it now parses integers and floats for the value. There is also now a fallback/error message for when a specified entry is not a `map[string]interface {}`.