Updated message_link.yag UNSTABLE RELEASE
oh boi... this was pain to code. Sooooo, you remember that discord set the message character limit higher from 2000 to 4000 for nitro (not classic) users... yea. This is a not very sexy fix that tries to add support for those huge 4000 character messages.

It is highly unstable but it works. If you find any bugs please report them right away.
Some of the layout of the embeds might be a little but messed up but hey... it works so i won't complain.