Updated the 2 'bypass limits' scripts
You can only use execCC 10 times per channel per minute.
Another fact you have to know about execCC is that the time is not really precise if you pick a small delay 8s - 1s. We could go as low as 6 seconds delay but that would cause issues if the 6 seconds take less once because it is not precise. Thats why i am using 8 to make sure it is 100% reliable & also as fast as possible