Updated 'Auto_Moderation/Anti_Phishing' - Added more domains
I added a lot more domains that got identified as phishing domains.
The biggest "issue" right now is not that the code doesn't work, but that there are just so many domains being created. I'd need to write some regex to compress the lists down into a smaller list, but I do not have the energy to do that.
For now this will have to do.