Bugfix for banlos of the logging cc
This has been an issue for way too long. I already deployed a patch for it but messed it up a while ago lol. Well good thing i did the patch because i can just use that to revert the db entries from "bans123456" to "ban123456" again when -bans is used xd